refactor(stt): relocate Whisper helpers from core/ports to lyra/nats/stt_helpers
Addresses #1224 review (architect): _WHISPER_NOISE_TOKENS, is_whisper_noise,
mime_from_suffix are adapter-specific (Whisper model + MIME table) not domain
port concerns. Aligns ports/stt.py with the ports/llm.py reference layout
(protocol + value object + error only).
