+++ src/lyra/core/CLAUDE.md
@@ -24,8 +24,9 @@ Adapters and infrastructure may import `core/`; never the reverse.
 
 `ports/llm.py`, `ports/stt.py`, `ports/tts.py` — pure Protocol definitions, no infrastructure imports.
 These are the hexagonal boundary: core declares what it needs; implementations live in `llm/drivers/` and `adapters/`.
-`ports/stt.py` and `ports/tts.py` follow the same pattern as `ports/llm.py`: protocol + value objects + errors + helpers, all co-located.
-The former `lyra.stt` and `lyra.tts` packages were deleted in #1221; `ports/` is the single owner.
+`ports/stt.py` and `ports/tts.py` follow the same pattern as `ports/llm.py`: protocol + value objects + errors only.
+The former `lyra.stt` and `lyra.tts` packages were deleted in #1221; `ports/` is the single owner of domain types.
+Adapter-adjacent helpers (`is_whisper_noise`, `mime_from_suffix`) live in `lyra/nats/stt_helpers.py`, not in `ports/`.
 
 ## Store pattern (ADR-048)
