+++ src/lyra/core/ports/stt.py
@@ -1,7 +1,8 @@
 """STTProtocol — Domain port for speech-to-text transcription.
 
-Self-contained: protocol + value object + errors + helpers.
-Pure: stdlib + pydantic only. No inbound lyra imports.
+Self-contained: protocol + value object + errors.
+Pure: stdlib only. No inbound lyra imports.
+Adapter-adjacent helpers (noise detection, MIME mapping) live in lyra.nats.stt_helpers.
 """
 
 from __future__ import annotations
@@ -34,37 +35,9 @@ class STTNoiseError(Exception):
     """
 
 
-_WHISPER_NOISE_TOKENS = {"[music]", "[applause]", "[laughter]", "[silence]", "[noise]"}
-
-
-def is_whisper_noise(text: str) -> bool:
-    """Return True if the text is empty or a known Whisper noise token."""
-    stripped = text.strip().lower()
-    return not stripped or stripped in _WHISPER_NOISE_TOKENS
-
-
-def mime_from_suffix(suffix: str) -> str:
-    """Map a file extension (with leading dot) to its audio MIME type.
-
-    Callers that receive audio as a file path (e.g. attachment handlers) use this
-    to derive the MIME type before calling STTProtocol.transcribe(audio, mime).
-    """
-    return {
-        ".ogg": "audio/ogg",
-        ".mp3": "audio/mpeg",
-        ".wav": "audio/wav",
-        ".m4a": "audio/mp4",
-        ".webm": "audio/webm",
-        ".flac": "audio/flac",
-        ".opus": "audio/ogg",
-    }.get(suffix.lower(), "audio/ogg")
-
-
 __all__ = [
     "STTProtocol",
     "TranscriptionResult",
     "STTUnavailableError",
     "STTNoiseError",
-    "is_whisper_noise",
-    "mime_from_suffix",
 ]
